refactor bottom select size
