+++ lib/src/widgets/button/button_add_favorite.dart
@@ -2,8 +2,9 @@ import 'package:e_commerce/src/config/themes/my_colors.dart';
 import 'package:e_commerce/src/constants/my_icons.dart';
 import 'package:e_commerce/src/models/products_model.dart';
 import 'package:e_commerce/src/modules/favorites/logic/favorites_bloc.dart';
+import 'package:e_commerce/src/utils/enum/page_info.dart';
 import 'package:e_commerce/src/widgets/bottom_sheet/bottom_sheet.dart';
-import 'package:e_commerce/src/widgets/bottom_sheet/bottom_sheet_size.dart';
+import 'package:e_commerce/src/widgets/bottom_sheet/bottom_sheet_select_size.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
@@ -45,7 +46,11 @@ class XButtonAddToFavorite extends StatelessWidget {
                     context.read<FavoriteBloc>().initSizeType();
 
                     XBottomSheet.show(context,
-                        widget: XBottomSheetFavorite(data: data));
+                        backgroundColor: MyColors.colorBackground,
+                        widget: XBottomSheetSelectSize(
+                          data: data,
+                          pageInfo: PageInfo.favorites,
+                        ));
                   },
                   child: const Image(
                     image: AssetImage(MyIcons.favoriteIcon),
